chore: Preview 배경 색상의 불필요한 타입 변환 연산자 제거
불필요한 toLong() 제거

+++ android/Staccato_AN/app/src/main/java/com/on/staccato/presentation/invitation/component/InvitationManagement.kt
@@ -59,10 +59,7 @@ fun InvitationManagement(
     }
 }
 
-@Preview(
-    showBackground = true,
-    backgroundColor = 0xFFFFFF.toLong(),
-)
+@Preview(showBackground = true, backgroundColor = 0xFFFFFF)
 @Composable
 private fun InvitationManagementPreview(
     @PreviewParameter(InvitationSelectionMenuProvider::class) menu: InvitationSelectionMenuUiModel,
